pages/requirements.txt
streamlit
pandas
numpy
scikit-learn
shap
plotly
mysql-connector-python